+++ NEWS.md
@@ -1,5 +1,8 @@
 # usethis 0.0.0.9000
 
+* `use_package_doc()` uses more modern roxygen2 template requires even less
+  duplication.
+
 * Removed old `add_build_ignore()`
 
 * `use_build_ignore()` now strips trailing `/`